Split TextUi class to Formatter class to adhere to SRP
